Split ExplorerWallet out of OIP
