+++ src/modules/wallets/index.js
@@ -1,0 +1,7 @@
+import ExplorerWallet from './ExplorerWallet'
+import RPCWallet from './RPCWallet'
+
+export {
+	ExplorerWallet,
+	RPCWallet
+}
